refactor(scripts): split validate suite into per-block modules + run.mjs validate
