+++ .github/workflows/ci.yml
@@ -18,4 +18,4 @@ jobs:
           node-version: 22
           cache: pnpm
       - run: pnpm install --frozen-lockfile
-      - run: node scripts/ci-validate.mjs
+      - run: node scripts/run.mjs validate
