+++ scripts/run.mjs
@@ -7,8 +7,9 @@
  * Wired so far:
  *   emit       — regenerate unified first-party manifests (write mode)
  *   emit-check — verify emitted products are fresh (drift → exit 1)
+ *   validate   — run the full 13-block validate suite
  *
- * Remaining commands (validate/version/publish-vendor/bump-submodule/
+ * Remaining commands (version/publish-vendor/bump-submodule/
  * apply-rules/smoke-cdd) are wired in by later tasks.
  */
 
@@ -22,13 +23,18 @@ const cmd = (name, desc, fn) =>
     .command(name)
     .description(desc)
     .action(async () => {
-      await import(fn).then((m) => m.main());
+      const code = await import(fn).then((m) => m.main());
+      // A numeric return is an exit code (validate main() → 1 on step failure);
+      // undefined returners (emit) rely on the top-level catch for non-zero.
+      if (typeof code === "number") process.exitCode = code;
     });
 
 // emit / emit-check wired first; the remaining commands join per-task in
-// subsequent tasks (note the commander subcommand name `emit-check`).
+// subsequent tasks (note the commander subcommand name `emit-check` and the
+// `validate` dispatcher in ./validate/index.mjs).
 cmd("emit", "regenerate unified first-party manifests", "./emit/all.mjs");
 cmd("emit-check", "verify emitted products are fresh (drift → exit 1)", "./emit/check.mjs");
+cmd("validate", "run the full validate suite (13 blocks)", "./validate/index.mjs");
 
 program.parseAsync(process.argv).catch((e) => {
   console.error(e.message);
